pcal/requirements.txt
img2pdf==0.5.1
pillow==11.0.0
pylint==3.3.2
pypdf==5.1.0
